Se agrego condicional para evitar salto de url sin sesion de usuario

+++ app/Http/Controllers/ProductsController.php
@@ -12,17 +12,17 @@ use App\ContainerModel;
 use App\User;
 
 class ProductsController extends Controller {
-    
-    public function index() {
 
-//        $search = Input::get('search');
-//        $btn_filter = 0;
-//  , 'btn_filter' => $btn_filter
-        $query = ProductsModel::select('id', 'sku', 'name', 'description')->where('status', '!=', '0')->orderBy('id','desc');
-        
-        $products = $query->orderBy('id')->paginate(3);
-    
-        return view('products/index', ['products' => $products]);
+    public function index() {
+        //Auth::check()
+        if (!Auth::guest()) {
+            $query = ProductsModel::select('id', 'sku', 'name', 'description', 'container_id')->where('status', '!=', '0');
+            $products = $query->orderBy('id', 'desc')->paginate(5);
+
+            return view('products/index', ['products' => $products]);
+        } else {
+            return redirect('/')->with('error', 'No tienes permiso para realizar esta acción. Intenta iniciando sesi&oacute;n');
+        }
     }
 
     /**
@@ -31,7 +31,11 @@ class ProductsController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function create() {
-        return view('products.create');
+        if (!Auth::guest()) {
+            return view('products.create');
+        } else {
+            return redirect('/')->with('error', 'No tienes permiso para realizar esta acción. Intenta iniciando sesi&oacute;n');
+        }
     }
 
     /**
@@ -41,83 +45,98 @@ class ProductsController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function store(ProductsRequest $request) {
+        if (!Auth::guest()) {
+            $producto = new ProductsModel();
+            $producto->sku = ucfirst(trim(preg_replace('/( ){2,}/u', ' ', $request->sku))); //1ra en mayuscula, para evitar espacios en blanco y borrarlos
+            $producto->name = ucfirst(trim(preg_replace('/( ){2,}/u', ' ', $request->name)));
+            $producto->description = ucfirst(trim(preg_replace('/( ){2,}/u', ' ', $request->description)));
+            $producto->status = 1;
+            $producto->container_id = trim(preg_replace('/( ){2,}/u', ' ', $this->setContainer()));
 
-        $producto = new ProductsModel();
-        $producto->sku = trim(preg_replace('/( ){2,}/u', ' ', $request->sku)); //1ra en mayuscula, para evitar espacios en blanco
-        $producto->name = trim(preg_replace('/( ){2,}/u', ' ', $request->name));
-        $producto->description = trim(preg_replace('/( ){2,}/u', ' ', $request->description));
-        $producto->status = 1;
-        $producto->container_id = trim(preg_replace('/( ){2,}/u', ' ', $this->setContainer()));
-        
-        $producto->save();
-        
-        if ($producto->save()) {
-            return redirect()->action('ProductsController@index')->with('exito', "Registro realizado con éxito.");
-        }else {
-            return redirect()->action('ProductsController@index')->with('error', "No se pudo realizar el registro.");
+            $producto->save();
+
+            if ($producto->save()) {
+                return redirect()->action('ProductsController@index')->with('exito', "Registro realizado con éxito.");
+            } else {
+                return redirect()->action('ProductsController@index')->with('error', "No se pudo realizar el registro.");
+            }
+        } else {
+            return redirect('/')->with('error', 'No tienes permiso para realizar esta acción. Intenta iniciando sesi&oacute;n');
         }
-
     }
 
 //setContainer
     public function setContainer() {
-        $contenedor = ContainerModel::findOrFail($this->getContainer());
-        if ($contenedor->capacity == 1000 && $contenedor->status == 1) {
-            $contenedor->status = 0;
-            $contenedor->ubication_id = 2;
-            $contenedor->save();
-            
-            $new_container = $this->getContainer();
+        if (!Auth::guest()) {
+            $contenedor = ContainerModel::findOrFail($this->getContainer());
+            if ($contenedor->capacity == 1000 && $contenedor->status == 1) {
+                $contenedor->status = 0;
+                $contenedor->ubication_id = 2;
+                $contenedor->save();
+
+                $new_container = $this->getContainer();
+            } else {
+                $contenedor->capacity = $contenedor->capacity + 1;
+                $contenedor->save();
+
+                $description = new DescriptionContainerModel();
+                $description->origin = 1;
+                $description->destinity = 2;
+                $description->status = 1;
+                $description->container_id = $contenedor->id;
+                $description->user_id = Auth::user()->id; //
+
+                $description->save();
+
+                $new_container = $this->getContainer();
+            }
+            return $new_container;
         } else {
-            $contenedor->capacity = $contenedor->capacity + 1;
-            $contenedor->save();
-            
-            $description = new DescriptionContainerModel();
-            $description->origin = 1;
-            $description->destinity = 2;
-            $description->status = 1;
-            $description->container_id = $contenedor->id;
-            $description->user_id = Auth::user()->id; //
-            
-            $description->save();
-            
-            $new_container = $this->getContainer();
+            return redirect('/')->with('error', 'No tienes permiso para realizar esta acción. Intenta iniciando sesi&oacute;n');
         }
-        return $new_container;
     }
 
     //getContainer
     public function getContainer() {
-        $containers = ContainerModel::select('id')->where('status', '=', 1)->orderBy('capacity', 'asc')->get();
-        if (count($containers) <= 0) {
-            $contenedor = new ContainerModel();
-            $contenedor->capacity = 0;
-            $contenedor->type = $this->getTypeContainer();
-            $contenedor->status = 1;
-            $contenedor->ubication_id = 1;
-            $contenedor->save();
+        if (!Auth::guest()) {
+            $containers = ContainerModel::select('id')->where('status', '=', 1)->orderBy('capacity', 'asc')->get();
+            if (count($containers) <= 0) {
+                $contenedor = new ContainerModel();
+                $contenedor->capacity = 0;
+                $contenedor->type = $this->getTypeContainer();
+                $contenedor->status = 1;
+                $contenedor->ubication_id = 1;
+                $contenedor->save();
 
-            $new_container = ContainerModel::where('capacity', '=', 0)->first();
-            $description = new DescriptionContainerModel();
-            $description->origin = 1;
-            $description->destinity = 2;
-            $description->status = 1;
-            $description->container_id = $new_container->id;
-            $description->user_id = Auth::user()->id; // sse guardará con el id del usuario autentificado
-            
-            $description->save();
-        } else {
-            foreach ($containers as $container) {
-                $new = $container->id;
+                $new_container = ContainerModel::where('capacity', '=', 0)->first();
+                $description = new DescriptionContainerModel();
+                $description->origin = 1;
+                $description->destinity = 2;
+                $description->status = 1;
+                $description->container_id = $new_container->id;
+                $description->user_id = Auth::user()->id; // sse guardará con el id del usuario autentificado
+
+                $description->save();
+            } else {
+                foreach ($containers as $container) {
+                    $new = $container->id;
+                }
+                $new_container = $new;
             }
-            $new_container = $new;
+            return $new_container;
+        } else {
+            return redirect('/')->with('error', 'No tienes permiso para realizar esta acción. Intenta iniciando sesi&oacute;n');
         }
-        return $new_container;
     }
 
     public function getTypeContainer() {
-        $type = mt_rand(1, 2);
-        return $type;
+        if (!Auth::guest()) {
+
+            $type = mt_rand(1, 2);
+            return $type;
+        } else {
+            return redirect('/')->with('error', 'No tienes permiso para realizar esta acción. Intenta iniciando sesi&oacute;n');
+        }
     }
 
     /**
@@ -137,10 +156,13 @@ class ProductsController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function edit($id) {
-        //la informacion se manda por url por ese motivo se utiliza get
-        $producto = ProductsModel::findOrFail($id);
-        //die(json_encode($producto));
-        return view('products.update', ['product' => $producto]);
+        if (!Auth::guest()) {
+            $producto = ProductsModel::findOrFail($id);
+            //die(json_encode($producto));
+            return view('products.update', ['product' => $producto]);
+        } else {
+            return redirect('/')->with('error', 'No tienes permiso para realizar esta acción. Intenta iniciando sesi&oacute;n');
+        }
     }
 
     /**
@@ -151,14 +173,19 @@ class ProductsController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id) {
-        $producto = ProductsModel::findOrFail($id); //realizamos la busqueda de los registros
-        //recibimos los valores nuevos 
-        $producto->sku = $request->sku;
-        $producto->name = $request->name;
-        $producto->description = $request->description;
-        //guardamos los cambios
-        $producto->save();
-        return redirect()->action('ProductsController@index')->with('actualizado', "Registro actualizado con éxito.");
+        if (!Auth::guest()) {
+            $producto = ProductsModel::findOrFail($id);
+
+            $producto->sku = $request->sku;
+            $producto->name = $request->name;
+            $producto->description = $request->description;
+
+            $producto->save();
+            
+            return redirect()->action('ProductsController@index')->with('actualizado', "Registro actualizado con éxito.");
+        } else {
+            return redirect('/')->with('error', 'No tienes permiso para realizar esta acción. Intenta iniciando sesi&oacute;n');
+        }
     }
 
     /**
@@ -168,8 +195,11 @@ class ProductsController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function destroy($id) {
-        Productos::findOrFail($id)->delete();
-        return redirect()->action('ProductosController@index');
+        if (!Auth::guest()) {
+            Productos::findOrFail($id)->delete();
+            return redirect()->action('ProductosController@index');
+        } else {
+            return redirect('/')->with('error', 'No tienes permiso para realizar esta acción. Intenta iniciando sesi&oacute;n');
+        }
     }
-
 }
